+++ appdaemon/adapi.py
@@ -1128,12 +1128,18 @@ class ADAPI:
     #
 
     @utils.sync_wrapper
-    async def register_endpoint(self, callback, name=None):
+    async def register_endpoint(
+        self, callback: Callable[[Any, dict], Any], endpoint: str = None, **kwargs: Optional[dict]
+    ) -> str:
         """Registers an endpoint for API calls into the current App.
 
         Args:
             callback: The function to be called when a request is made to the named endpoint.
-            name (str, optional): The name of the endpoint to be used for the call  (Default: ``None``).
+            endpoint (str, optional): The name of the endpoint to be used for the call  (Default: ``None``).
+            This must be unique across all endpoints, and when not given, the name of the app is used as the endpoint.
+            It is possible to register multiple endpoints to a single app instance.
+        Keyword Args:
+            **kwargs (optional): Zero or more keyword arguments.
 
         Returns:
             A handle that can be used to remove the registration.
@@ -1143,21 +1149,25 @@ class ADAPI:
             and an HTTP OK status response (e.g., `200`. If this is not added as a returned response,
             the function will generate an error each time it is processed.
 
-            >>> self.register_endpoint(my_callback)
-            >>> self.register_callback(alexa_cb, "alexa")
+            >>> self.register_endpoint(self.my_callback)
+            >>> self.register_endpoint(self.alexa_cb, "alexa")
+
+            >>> async def alexa_cb(self, request, kwargs):
+            >>>     data = await request.json()
+            >>>     self.log(data)
+            >>>     response = {"message": "Hello World"}
+            >>>     return response, 200
 
         """
-        if name is None:
-            ep = self.name
-        else:
-            ep = name
+        if endpoint is None:
+            endpoint = self.name
+
         if self.AD.http is not None:
-            return await self.AD.http.register_endpoint(callback, ep)
+            return await self.AD.http.register_endpoint(callback, endpoint, self.name, **kwargs)
         else:
             self.logger.warning(
-                "register_endpoint for %s filed - HTTP component is not configured", name,
+                "register_endpoint for %s failed - HTTP component is not configured", endpoint,
             )
-            return None
 
     @utils.sync_wrapper
     async def deregister_endpoint(self, handle: str) -> None:
@@ -1180,7 +1190,9 @@ class ADAPI:
     #
 
     @utils.sync_wrapper
-    async def register_route(self, callback, route=None, **kwargs):
+    async def register_route(
+        self, callback: Callable[[Any, dict], Any], route: str = None, **kwargs: Optional[dict]
+    ) -> str:
         """Registers a route for Web requests into the current App.
            By registering an app web route, this allows to make use of AD's internal web server to serve
            web clients. All routes registered using this api call, can be accessed using
@@ -1191,10 +1203,7 @@ class ADAPI:
             route (str, optional): The name of the route to be used for the request (Default: the app's name).
 
         Keyword Args:
-            token (str, optional): A previously registered token can be passed with the api call, which
-            can be used to secure the app route. This allows for different security credentials to be used across different
-            app routes. It should be noted that if a device has already registered using AD's Admin UI's password
-            and a cookie has been stored by the browser, that device will bypass the token and still access the web server.
+            **kwargs (optional): Zero or more keyword arguments.
 
         Returns:
             A handle that can be used to remove the registration.
@@ -1216,7 +1225,6 @@ class ADAPI:
 
         else:
             self.logger.warning("register_route for %s filed - HTTP component is not configured", route)
-            return None
 
     @utils.sync_wrapper
     async def deregister_route(self, handle: str) -> None: